+++ src/main/java/main/controllers/Project/SuiteController.java
@@ -3,24 +3,28 @@ package main.controllers.Project;
 import main.controllers.BaseController;
 import main.exceptions.AqualityException;
 import main.exceptions.AqualityPermissionsException;
-import main.model.db.dao.project.SuiteStatisticDao;
-import main.model.db.dao.project.TestSuiteDao;
-import main.model.dto.SuiteStatisticDto;
-import main.model.dto.TestDto;
-import main.model.dto.TestSuiteDto;
-import main.model.dto.UserDto;
+import main.model.db.dao.project.*;
+import main.model.dto.*;
+import main.view.Project.TestResult;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class SuiteController extends BaseController<TestSuiteDto> {
     private TestSuiteDao testSuiteDao;
+    private TestRunDao testRunDao;
+    private Test2SuiteDao test2SuiteDao;
+    private TestResultDao testResultDao;
     private SuiteStatisticDao suiteStatisticDao;
     private TestController testController;
 
     public SuiteController(UserDto user) {
         super(user);
         testSuiteDao = new TestSuiteDao();
+        testRunDao = new TestRunDao();
+        testResultDao = new TestResultDao();
         suiteStatisticDao = new SuiteStatisticDao();
+        test2SuiteDao = new Test2SuiteDao();
         testController = new TestController(user);
     }
 
@@ -74,6 +78,51 @@ public class SuiteController extends BaseController<TestSuiteDto> {
         }
     }
 
+    public List<TestDto> findLegacyTests(Integer suiteId, Integer notExecutedFor) throws AqualityException {
+        if (baseUser.isManager() || baseUser.getProjectUserBySuiteId(suiteId).isEditor()) {
+            TestRunDto testRunTemplate = new TestRunDto();
+            testRunTemplate.setTest_suite_id(suiteId);
+            testRunTemplate.setLimit(notExecutedFor);
+            List<TestRunDto> testRuns = testRunDao.searchAll(testRunTemplate);
+            TestDto testTemplate = new TestDto();
+            testTemplate.setTest_suite_id(suiteId);
+            List<TestDto> tests = testController.get(testTemplate);
+            for (TestRunDto testRun : testRuns) {
+                TestResultDto testResultTemplate = new TestResultDto();
+                testResultTemplate.setTest_run_id(testRun.getId());
+                List<TestResultDto> testResults = testResultDao.searchAll(testResultTemplate);
+                tests = tests.stream().filter(test -> {
+                    TestResultDto currentResult = testResults.stream().filter(result -> result.getTest_id().equals(test.getId())).findFirst().orElse(null);
+                    return currentResult == null || currentResult.getFinal_result_id() == 3;
+                }).collect(Collectors.toList());
+            }
+
+            return tests;
+        } else {
+            throw new AqualityPermissionsException("Account is not allowed to Sync Test Suite", baseUser);
+        }
+    }
+
+    public void syncLegacyTests(List<TestDto> legacyTests, Integer suiteId, boolean removeNotExecutedResults) throws AqualityException {
+        if(baseUser.isManager() || baseUser.getProjectUserBySuiteId(suiteId).isEditor()){
+            legacyTests.forEach(test -> {
+                Test2SuiteDto test2Suite = new Test2SuiteDto();
+                test2Suite.setTest_id(test.getId());
+                test2Suite.setSuite_id(suiteId);
+                try {
+                    test2SuiteDao.delete(test2Suite);
+                    if(removeNotExecutedResults) {
+                        removePendingResultsForTest(suiteId, test.getId());
+                    }
+                } catch (AqualityException e) {
+                    e.printStackTrace();
+                }
+            });
+        }else{
+            throw new AqualityPermissionsException("Account is not allowed to Sync Test Suite", baseUser);
+        }
+    }
+
     private List<TestSuiteDto> fillTestSuites(List<TestSuiteDto> testSuites, boolean withChildren) throws AqualityException {
         if(withChildren){
             for (TestSuiteDto suite: testSuites){
@@ -85,5 +134,16 @@ public class SuiteController extends BaseController<TestSuiteDto> {
             }
         }
         return testSuites;
+    }
+
+    private void removePendingResultsForTest(Integer suiteId, Integer testId) throws AqualityException {
+        List<TestResultDto> testResults = testResultDao.selectSuiteLegacyResults(suiteId, testId);
+        testResults.forEach(testResult -> {
+            try {
+                testResultDao.delete(testResult);
+            } catch (AqualityException e) {
+                e.printStackTrace();
+            }
+        });
     }
 }